feat: introduce new features in trasactions page
transactions page now fully communicates with it's backend, being able
to fetch, delete, create and update all teh transactions.

Transactions are also sorted based on users choice of preference and
only transactions related to chosen month are display.

There is quite some data formatting needed in order to achieve the
wanted state which probably will be handled on backend in the future as
it probably will be much  better performace-wise

also modal was removed and one simple form was created directly on the
page and the graph was seen unneccessary and removed.